// src/app/routes.enum.ts
export enum RoutesEnum {
  LOGIN = 'login',
  REGISTER = 'register',
  SHOPS = 'shops',
  CARD = 'card',
  SHIPPING = 'shipping',
  PAYMENT = 'payment',
  PRODUCTS = 'products',
  PRODUCT = 'product',
  SEARCH = 'search',
  DETAILS = 'details',
  ORDER = 'order',
  CONFIRM = 'confirm',
  DASHBOARD = 'dashboard',
  MY_ORDERS = 'my-orders',
  CHANGE_PASSWORD = 'change-password',
  MY_WHISHLIST = 'my-wishlist',
  CHAT = 'chat',
}